requirements.txt
streamlit
pandas
plotly
plotly-express
torch
torchvision
torchaudio
ultralytics
opencv-python-headless
Pillow
scikit-learn
matplotlib
transformers
clip-openai
supervision
numpy
